trading-ia-bot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
